Separated object 'status' into object 'obj_status' and player 'pl_status', both of which are 16 bits, but only the former is in the object structure. This will potentially save 2 bytes per object. Added inline function bool Player_is_paused(player_t *).

+++ xpilot/src/server/netserver.h
@@ -44,7 +44,7 @@ int Send_self(connection_t *connp, player_t *pl,
 	      int lock_dist,
 	      int lock_dir,
 	      int autopilotlight,
-	      long status,
+	      int status,
 	      char *mods);
 int Send_leave(connection_t *connp, int id);
 int Send_player(connection_t *connp, int id);
